requirements.txt
requests==2.31.0
vapi_python
python-dotenv==1.0.0
fastapi
uvicorn
pydantic
httpx
